+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.75.0] - 2019-05-20
+### Changed
+- Use better resolver architecture at productSearch query.
+
+### Added
+- breadcrum resolver at productSearch.
+
 ## [2.74.4] - 2019-05-20
 
 ### Fixed